requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
scikit-learn==1.3.1
joblib==1.3.2
hydra-core==1.3.2
pydantic==2.4.2
numpy==1.26.0
lightgbm==4.1.0
pyyaml==6.0.1